requirements.txt
Django==2.2
django-debug-toolbar==3.1.1
django-softdelete==0.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1